Tests/LibWeb/Text/expected/abspos-relayout-remove-reinsert.txt

moving=55,40,35,20
partial-layouts=0 full-layouts=1
FAIL
